+++ lib/block/blocktransfer-error.js
@@ -12,7 +12,7 @@
 
 'use strict';
 
-const errorCodes = require('./error-codes');
+const errorCodes = require('../http-error-codes');
 
 /**
  * Custom Error class containing additional information specific to the download process. This primarily consists of an
