+++ lib/block/download-error.js
@@ -12,7 +12,7 @@
 
 'use strict';
 
-const errorCodes = require('./error-codes');
+const errorCodes = require('../http-error-codes');
 const BlockTransferError = require('./blocktransfer-error');
 
 /**
