+++ lib/functions/aeminitiateupload.js
@@ -18,7 +18,7 @@ const { AssetMetadata } = require("../asset/assetmetadata");
 const { AssetMultipart } = require("../asset/assetmultipart");
 const { AsyncGeneratorFunction } = require("../generator/function");
 const UploadError = require("../block/upload-error");
-const ErrorCodes = require("../block/error-codes");
+const ErrorCodes = require("../http-error-codes");
 const { postForm } = require("../fetch");
 const { retry } = require("../retry");
 const logger = require("../logger");
